+++ livekit-protocol/CHANGELOG.md
@@ -62,6 +62,17 @@
 ### Changed
 
 - Update protocol version to v1.31.0
+## 0.7.5 (2026-04-23)
+
+### Features
+
+#### Add support for frame level packet trailer
+
+##890 by @chenosaurus
+
+- Add support to attach/parse frame level timestamps & frame ID to VideoTracks as a custom payload trailer.
+- Breaking change in VideoFrame API, must include `frame_metadata` or use VideoFrame::new().
+
 ## 0.7.4 (2026-04-02)
 
 ### Fixes